revised documentation in sukarno speech notebook and related files. renamed sentiment analysis folder to text analysis
