+++ text_analysis/data/books.sframe/m_17a829304bd6788e.frame_idx
@@ -1,0 +1,13 @@
+[sframe]
+version=0
+num_segments=0
+num_columns=3
+nrows=64
+[column_names]
+0000=link
+0001=text
+0002=title
+[column_files]
+0000=m_17a829304bd6788e.sidx:0
+0001=m_17a829304bd6788e.sidx:1
+0002=m_17a829304bd6788e.sidx:2
