+++ text_analysis/data/sentences.sframe/m_256796f87804b131.frame_idx
@@ -1,0 +1,13 @@
+[sframe]
+version=0
+num_segments=0
+num_columns=3
+nrows=70971
+[column_names]
+0000=title
+0001=sentence
+0002=words
+[column_files]
+0000=m_256796f87804b131.sidx:0
+0001=m_256796f87804b131.sidx:1
+0002=m_256796f87804b131.sidx:2
