+++ text_analysis/data/sentences.sframe/m_256796f87804b131.sidx
@@ -1,0 +1,38 @@
+{
+	"sarray" : {
+		"version" : 2,
+		"num_segments" : 1
+	},
+	"segment_files" : {
+		"0000" : "m_256796f87804b131.0000"
+	},
+	"columns" : [
+		{
+			"content_type" : "",
+			"metadata" : {
+				"__type__" : "2"
+			},
+			"segment_sizes" : {
+				"0000" : "70971"
+			}
+		},
+		{
+			"content_type" : "",
+			"metadata" : {
+				"__type__" : "2"
+			},
+			"segment_sizes" : {
+				"0000" : "70971"
+			}
+		},
+		{
+			"content_type" : "",
+			"metadata" : {
+				"__type__" : "4"
+			},
+			"segment_sizes" : {
+				"0000" : "70971"
+			}
+		}
+	]
+}
