+++ text_analysis/data/sukarno.sframe/dir_archive.ini
@@ -1,0 +1,9 @@
+[archive]
+version=1
+num_prefixes=3
+[metadata]
+contents=sframe
+[prefixes]
+0000=dir_archive.ini
+0001=objects.bin
+0002=m_ee5ab27d0550b39d
