+++ text_analysis/data/sukarno.sframe/m_ee5ab27d0550b39d.sidx
@@ -1,0 +1,20 @@
+{
+	"sarray" : {
+		"version" : 2,
+		"num_segments" : 1
+	},
+	"segment_files" : {
+		"0000" : "m_ee5ab27d0550b39d.0000"
+	},
+	"columns" : [
+		{
+			"content_type" : "",
+			"metadata" : {
+				"__type__" : "2"
+			},
+			"segment_sizes" : {
+				"0000" : "1"
+			}
+		}
+	]
+}
