+++ text_analysis/data/topic_model/dir_archive.ini
@@ -1,0 +1,8 @@
+[archive]
+version=1
+num_prefixes=2
+[metadata]
+contents=model
+[prefixes]
+0000=dir_archive.ini
+0001=objects.bin
